Move most of sectors aggregates to a component and generalise away fro msectors
